renderloop update scene loading now initing lights right

+++ src/scene/PointLight.hpp
@@ -5,7 +5,7 @@
 #include "Light.hpp"
 
 struct PointLightParams {
-    uint32_t index;
+    size_t index;
     float constant = 1.0f;
     float linear = 0.35f;
     float quadratic = 0.44f;
@@ -27,7 +27,7 @@ private:
     float linear;
     float quadratic;
 
-    uint32_t index;
+    size_t index;
     
 };
